+++ src/jfan/fan/sys/Depend.java
@@ -177,7 +177,7 @@ public final class Depend
     return toStr().hashCode();
   }
 
-  public Long hash()
+  public long hash()
   {
     return FanStr.hash(toStr());
   }
@@ -215,33 +215,33 @@ public final class Depend
     return name;
   }
 
-  public final Long size()
+  public final long size()
   {
-    return Long.valueOf(constraints.length);
+    return constraints.length;
   }
 
   public final Version version() { return version(0L); }
-  public final Version version(Long index)
+  public final Version version(long index)
   {
-    return constraints[index.intValue()].version;
+    return constraints[(int)index].version;
   }
 
   public final boolean isPlus() { return isPlus(0L); }
-  public final boolean isPlus(Long index)
+  public final boolean isPlus(long index)
   {
-    return constraints[index.intValue()].isPlus;
+    return constraints[(int)index].isPlus;
   }
 
   public final boolean isRange() { return isRange(0L); }
-  public final boolean isRange(Long index)
+  public final boolean isRange(long index)
   {
-    return constraints[index.intValue()].endVersion != null;
+    return constraints[(int)index].endVersion != null;
   }
 
   public final Version endVersion() { return endVersion(0L); }
-  public final Version endVersion(Long index)
+  public final Version endVersion(long index)
   {
-    return constraints[index.intValue()].endVersion;
+    return constraints[(int)index].endVersion;
   }
 
   public final boolean match(Version v)
@@ -252,14 +252,14 @@ public final class Depend
       if (c.isPlus)
       {
         // versionPlus
-        if (c.version.compare(v).longValue() <= 0)
+        if (c.version.compare(v) <= 0)
           return true;
       }
       else if (c.endVersion != null)
       {
         // versionRange
-        if (c.version.compare(v).longValue() <= 0 &&
-            (c.endVersion.compare(v).longValue() >= 0 || doMatch(c.endVersion, v)))
+        if (c.version.compare(v) <= 0 &&
+            (c.endVersion.compare(v) >= 0 || doMatch(c.endVersion, v)))
           return true;
       }
       else
